locli:  drop legacy code paths
