+++ bench/locli/src/Cardano/Util.hs
@@ -1,20 +1,23 @@
 module Cardano.Util
   ( module Cardano.Util
+  , module Control.Arrow
   , module Control.Concurrent.Async
+  , module Control.Monad.Trans.Except.Extra
   , module Text.Printf
   )
 where
 
-import Prelude                    (String)
+import Prelude                          (String)
 import Cardano.Prelude
 
-import Control.Concurrent.Async   (forConcurrently, forConcurrently_, mapConcurrently, mapConcurrently_)
-import Control.DeepSeq            qualified as DS
+import Control.Arrow                    ((&&&), (***))
+import Control.Concurrent.Async         (forConcurrently, forConcurrently_, mapConcurrently, mapConcurrently_)
+import Control.DeepSeq                  qualified as DS
 import Control.Monad.Trans.Except.Extra (firstExceptT, newExceptT)
-import Data.Aeson                 (FromJSON, ToJSON, encode, eitherDecode)
-import Data.ByteString.Lazy.Char8 qualified as LBS
-import Data.Text                  qualified as T
-import Text.Printf                (printf)
+import Data.Aeson                       (FromJSON, ToJSON, encode, eitherDecode)
+import Data.ByteString.Lazy.Char8       qualified as LBS
+import Data.Text                        qualified as T
+import Text.Printf                      (printf)
 
 import System.FilePath                  qualified as F
